#1593 partial commit

+++ application/controllers/webservice/newhomewebservice.php
@@ -41,7 +41,7 @@ class NewHomeWebService extends MY_Controller
         $this->slugerrorjson = file_get_contents(APPPATH . "resources/json/slugerrorjson.json");        
         $this->json = file_get_contents(APPPATH . "resources/json/jsonp.json");    
 
-        if($this->input->get()) {
+        if($this->input->get()) {        
             $this->authentication($this->input->get(), $this->input->get('hash'));
         }    
     }
@@ -58,7 +58,9 @@ class NewHomeWebService extends MY_Controller
 
         $index = $index == 0 ? 1 : $index + 1;
         $subIndex = $subIndex == 0 ? 1 : $subIndex + 1;
-
+        if($nodename == "subSliderSection" || $nodename == "mainSliderSection") {
+            $this->file = $this->tempHomefile;
+        }
         $remove = $this->xmlCmsService->removeXmlNode($this->file,$nodename,$index, $subIndex);
         if($remove == true) {
             return $this->output
@@ -240,7 +242,20 @@ class NewHomeWebService extends MY_Controller
         $this->output
             ->set_content_type('text/plain') 
             ->set_output(file_get_contents($this->file));
-    }    
+    }   
+
+    /**
+     *  Method to display the contents of the home_files.xml from the function call from Easyshop.ph.admin
+     *
+     *  @return string
+     */
+    public function getSliderContents() 
+    {
+        $this->fetchPreviewSlider(true);
+        $this->output
+            ->set_content_type('text/plain') 
+            ->set_output(file_get_contents($this->tempHomefile));
+    }      
 
     /**
      *  Sets Brand Section
@@ -805,7 +820,7 @@ class NewHomeWebService extends MY_Controller
         $subIndex = (int)$this->input->get("subIndex");
         $target = $this->input->get("target");
         $value = $this->input->get("value");
-        $map = simplexml_load_file($this->file);        
+        $map = simplexml_load_file($this->tempHomefile);        
 
         if(!empty($_FILES['myfile']['name'])) {
             $filename = date('yhmdhs');
@@ -839,7 +854,7 @@ class NewHomeWebService extends MY_Controller
             $map->sliderSection->slide[$index]->image[$subIndex]->path = $map->sliderSection->slide[$index]->image[$subIndex]->path;
             $map->sliderSection->slide[$index]->image[$subIndex]->target = $target;
         }
-        if($map->asXML($this->file)) {
+        if($map->asXML($this->tempHomefile)) {
             return $this->output
                     ->set_content_type('application/json')
                     ->set_output($this->json);
@@ -855,14 +870,14 @@ class NewHomeWebService extends MY_Controller
      */
     public function setSliderDesignTemplate()
     {
-        $map = simplexml_load_file($this->file);
+        $map = simplexml_load_file($this->tempHomefile);
 
         $index = (int)$this->input->get("index");
         $value = $this->input->get("value");        
 
         $map->sliderSection->slide[$index]->template = $value;
 
-        if($map->asXML($this->file)) {
+        if($map->asXML($this->tempHomefile)) {
             return $this->output
                     ->set_content_type('application/json')
                     ->set_output($this->json);
@@ -877,7 +892,7 @@ class NewHomeWebService extends MY_Controller
     {
         $template = [];
         $image = [];
-        $map = simplexml_load_file($this->file);        
+        $map = simplexml_load_file($this->tempHomefile);        
         $order = (int) $this->input->get("order");  
         $index = (int)  $this->input->get("index");  
         $nodename =  $this->input->get("nodename");  
@@ -895,10 +910,10 @@ class NewHomeWebService extends MY_Controller
         foreach($map->sliderSection->slide[$sliderOrder]->image as $images) {
                 $image[] = $images;
         }
-        $this->xmlCmsService->removeXmlNode($this->file,$nodename,$sliderOrder + 1); 
+        $this->xmlCmsService->removeXmlNode($this->tempHomefile,$nodename,$sliderOrder + 1); 
         $string = $this->xmlCmsService->getString("sliderSection", $template, "", "", "");      
-        $this->xmlCmsService->addXmlFormatted($this->file,$string,'/map/sliderSection/slide['.($sliderOrder + 1).']',"\t\t","\n\n");
-        $this->xmlCmsService->syncSliderValues($this->file,$image,$template,$sliderIndex,$sliderOrder);
+        $this->xmlCmsService->addXmlFormatted($this->tempHomefile,$string,'/map/sliderSection/slide['.($sliderOrder + 1).']',"\t\t","\n\n");
+        $this->xmlCmsService->syncSliderValues($this->tempHomefile,$image,$template,$sliderIndex,$sliderOrder);
         return $this->output
                 ->set_content_type('application/json')
                 ->set_output($this->json);            
@@ -910,7 +925,7 @@ class NewHomeWebService extends MY_Controller
      */
     public function setSliderPosition()
     {
-        $map = simplexml_load_file($this->file);
+        $map = simplexml_load_file($this->tempHomefile);
         $order = (int) $this->input->get("order");  
         $index = (int)  $this->input->get("index");  
         $subIndex = (int) $this->input->get("subIndex"); 
@@ -924,7 +939,7 @@ class NewHomeWebService extends MY_Controller
         $map->sliderSection->slide[$index]->image[$subIndex]->path =  $tempPath;
         $map->sliderSection->slide[$index]->image[$subIndex]->target =  $tempTarget;
     
-        if($map->asXML($this->file)) {
+        if($map->asXML($this->tempHomefile)) {
             return $this->output
                     ->set_content_type('application/json')
                     ->set_output($this->json);
@@ -945,16 +960,21 @@ class NewHomeWebService extends MY_Controller
                 ->set_output($this->json);                
     }
 
-    public function fetchPreviewSlider()
+    public function fetchPreviewSlider($isForSync = false)
     {
         $map = simplexml_load_file($this->file);
 
         foreach ($map->sliderSection->slide as $key => $slider) {
             $sliders[] = $slider;
         }
-        $this->xmlCmsService->removeXmlNode($this->tempHomefile,"tempHomeSlider");
-        $this->xmlCmsService->syncTempSliderValues($this->tempHomefile,$this->file,$sliders);
+        if($this->input->get("userid") == NULL) {
+            $this->xmlCmsService->removeXmlNode($this->tempHomefile,"tempHomeSlider");
+            $this->xmlCmsService->syncTempSliderValues($this->tempHomefile,$this->file,$sliders);
+        }
 
+        if($isForSync) {
+            return true;
+        }
         $homeContent = $this->serviceContainer['xml_cms']->getHomeData(false, true);
 
         $sliderSection = $homeContent['slider']; 
@@ -980,7 +1000,7 @@ class NewHomeWebService extends MY_Controller
         $file_ext = explode('.', $_FILES['myfile']['name']);
         $file_ext = strtolower(end($file_ext));  
         $path_directory = 'assets/images/homeslider';
-        $map = simplexml_load_file($this->file);
+        $map = simplexml_load_file($this->tempHomefile);
         $this->upload->initialize(array( 
             "upload_path" => $path_directory,
             "overwrite" => FALSE, 
@@ -1003,7 +1023,7 @@ class NewHomeWebService extends MY_Controller
             if($map->sliderSection->slide[$index]->image->path == "unavailable_product_img.jpg" && $map->sliderSection->slide[$index]->image->target == "/") {
                 $map->sliderSection->slide[$index]->image->path = $value;
                 $map->sliderSection->slide[$index]->image->target = $target;
-                if($map->asXML($this->file)) {
+                if($map->asXML($this->tempHomefile)) {
                     return $this->output
                             ->set_content_type('application/json')
                             ->set_output($this->json);
@@ -1011,7 +1031,7 @@ class NewHomeWebService extends MY_Controller
             }
             else {
                 $index = $index == 0 ? 1 : $index + 1;
-                $addXml = $this->xmlCmsService->addXmlFormatted($this->file,$string,'/map/sliderSection/slide['.$index.']/image[last()]',"\t\t\t","\n");
+                $addXml = $this->xmlCmsService->addXmlFormatted($this->tempHomefile,$string,'/map/sliderSection/slide['.$index.']/image[last()]',"\t\t\t","\n");
 
             }
             if($addXml === TRUE) {
